+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   is a shared `Arc<AtomicBool>` which can be used to stop rendering.  The
   returned type is now an `Option<...>`, where `None` indicates that rendering
   was cancelled.
+- Fix inconsistency between JIT and VM evaluator when performing interval
+  evaluation of `not([NAN, NAN])`.
 
 # 0.3.4
 - Add `GenericVmFunction::simplify_with` to simultaneously simplify a function